requirements.txt
streamlit
PyPDF2
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
langchain-groq
python-dotenv
huggingface-hub
